scanom-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
pillow
numpy
python-multipart
scipy
scikit-fuzzy
networkx
google-generativeai
tensorflow
